jobs-framework-api: initial support for user validation
This patch introduces the initial support for user validation. This is a core
part of the whole Toolforge jobs framework.

The incoming request should contain the 'ssl-client-subject-dn' header, which
has been genereated by nginx when validating the client cert. With that header,
we directly try to load the .kube/config from that same user from its home
directory.

Bug: T283238
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I632309c4ed5ff267b4486d5b56ddf09dc726af22