jobs: cleanup pods created by jobs
Before this patch, deleting Jobs, CronJobs or Deployments would leave the associated pods dangling
in the cluster.

Per https://kubernetes.io/docs/concepts/workloads/controllers/garbage-collection/ we need to
specify the garbage collection policy to cascade delete pods when working with CronJobs and
Deployments.

For base Jobs (the k8s object) we can specify `ttlSecondsAfterFinished` to somewhat mirror the
grid behavior: once a job is done, it no longer appears in the grid, and is thus no longer
registered anywhere.
However, if base Jobs objects are deleted and the associated pod were still running, we need to
force deletion of those pods. Include for this case an explicit deletion of pods.

Bug: T283238
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Id62f544282e08db8050de03175f6cc78662fb041