jobs: encode object type in a label and use it
Certain operations (listing, deleting) may not properly ignore (or ignore) objects created
in kubernetes.

Before this patch, when listing all jobs, if a cronjob k8s object had created a job object
(expected) then we would list the job object as one of our jobs, which is wrong.

To detect this parent-child relationship (which is also very useful for delete operations) encode
the k8s type in a label.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I56345810d5254c5dcf800ff4b0802291f9307e66