k8sclient: Validate TLS, set user-agent
Given we have the CA certificate available inside the container, no
reason to not validate it. Webservice did not validate it due to some
Python 2 bugs that do not apply here.

Also set a meaningful User-Agent in the spirit of the UA policy[0] to
easily identify TJF requests in Kubernetes API server request logs.

[0]: https://meta.wikimedia.org/wiki/User-Agent_policy

Change-Id: Ieab3619fe7da9d1ff8887d1e682755e13ee1dac5