jobs: include option to log job stdout/stderr to the each user home directory
Users home directories are on NFS and are our current way to share storage between the k8s cluster
and bastions.

This patch introduces a new API argument: filelog (boolean). If set to true, we will force redirect
stdout/stderr to a file in the home dir with the name of the job. If set to false (the default)
stdout/stderr will be redirected to /dev/null.

For this to happen, we wrap the whole job command in a /bin/sh execution.
To simplify a bit the whole parsing game when listing back jobs, we store the log preference in a
custom k8s label.

Incidentally, this enables commands to be arbitrary strings, meaning we can have arguments for the
job command, for example:

   ./arg.sh one --2 'three four'

Bug: T286485
Bug: T286107
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I2321d5c7061135bdf47265373a4c8e8e5b945c6a