jobs-framework-api: add --retry to api
* --retry arg added
* added tests for retry
For now maximum retry limit is set to 5 to encourage fair and
responsible usage of the API. If this value is too low/ high,
it can be changed.
Note that to run the test you might need to specifically run the
"retried job" tests or else you might risk getting inconsistent
results. Tests right now are too inconsistent. This will probably
be addressed in a different patch though.

Bug: T315114
Bug: T304893
Change-Id: I6fa64f5e15038dcf21915e63de4bb95ce93313d1