deployment: refresh generic image, try to avoid the standalone image
The standalone image is meant to be used as the base for building other
images, but not directly.

Introduce a new image `tf-bullseye-std`, which uses the more sensible -sssd
variant and mark the standalone as deprecated.

Tools using the old -standalone image would see a -DEPRECATE tag when listing.
Tools using the old -std shortname will continue loading from a jobs.yaml file.

Bug: T321612
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ibb846cb60543bb3a09d238f3548495b557d1ab3d