command: introduce filelog_std{out,err} parameters
We plan to expose them to users in the API in a later patch.

This patch also reworks how the command array is passed (and retrieved) from
kubernetes.

Before this patch:
 /bin/sh
 -c
 --
 ./user-command.sh 1>>/dev/null 2>>/dev/null

With this patch:
 /bin/sh
 -c
 --
 exec 1>>/dev/null;exec 2>>/dev/null;./user-command.sh

To be able to parse back both formats (until old-format jobs go away) a new
metadata label is introduced. This allows for a robust parsing without playing
string parsing games.

Bug: T301901
Bug: T325755
Bug: T302211
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I613aa844987ee31086d3e04280b751ed324d1edc