job: Hardcode container name as 'job'
Kubernetes applies more strict limits to container names compared to
deployment or cronjob names. Since we don't actually use the job name
for anything currently we can simply hardcode it as 'job' to avoid those
restrictions. (This is similar to the webservice implementation, which
uses 'webservice' as the container name regardless of the name of the
tool or the webservice.)

Bug: T330776
Change-Id: I18b9b37cfbe06301ad6e05931307eb5ce82adea6