command: Do not void log output if filelog is disabled
Updates the command logging handling (again) to disable the output
redirection to /dev/null when file logging is disabled. There are two
separate reasons why not discarding the output is wanted:
 * First, buildpack image entrypoints are not compatible with the file
   log setup. So there needs to be a mechanism to disable that entirely.
 * Additionally, the Loki-based logging setups that I've been playing
   with would ingest data from the container logs too. This is not
   something that's happening soon, but I figured I'd mention it anyway.

The primary reason (AIUI) why this was done previously was to protect
the Kubernetes nodes disks from filling up. Modern versions of
Kubernetes however include automatic log rotation[0], and I tested this
with a container producing large log volumes[1], and the log file in
/var/lib/docker/containers was kept at 100M as expected.

[0]: https://kubernetes.io/docs/concepts/cluster-administration/logging/#log-rotation
[1]: https://phabricator.wikimedia.org/P47577

Bug: T330715
Change-Id: Ie43a7226603a6d1e4c0b299de95dbc4096cfea64